(4#) - Add main example about how to test with react-testing-library
This test can be useful as basic example of how to use our mock
axios intance and how to integrate it with a component. In this case,
the CreateDiet component.

This test includes also how to wait a given element to be rendered
into the DOM after a promise is resolved.